magentamen-picks/requirements.txt
Flask
python-dotenv
requests
SQLAlchemy 